Allow `TabBarBottom` accept `onTabPress` callback (#486)
Pressing a tab bar often requires performing some kind of side effects
other than navigating: _Scroll-to-top_ functionality, for instance.
At the moment `TabView.TabBarTop`, which is the default on _Android_,
handles `onTabPress` callback as expected, but `TabView.TabBarBottom`
ignores it.

This PR addresses the problem so that `TabView.TabBarBottom` can accept
and handle the optional `onTabPress` callback. The callback is invoked
exactly once if present on every tab bar press, accepting as a sole
parameter a `NavigationRoute`, which is the route object the pressed
target represents.